Fixes Octavia handling of subnets without DHCP
Currently Octavia assumes that DHCP service is available on
the VIP and member subnets.  This is not the case at all operators.
This patch makes Octavia use the IP information provided when
the ports are created, if available.  If the IP information is
not available on the ports it will fall back to relying on DHCP.

Change-Id: I08a93d4318bbce48128019376320782d1a334369
Closes-Bug: #1607900