Move bandit to pep8
The bandit gate is passing, let's make it voting.

Since bandit is a security linter, add it to the linter target - called
pep8 - and run it there. This will save the CI systems spawning an extra
test machine for this.

This is the recommended approach in OpenStack CI to run all linters as
part of pep8 target - see also the consistent testing interface for
python (http://governance.openstack.org/reference/cti/python_cti.html)
that associates pep8 with "Codestyle checks".

Change-Id: Id65a50c35723b79e9b92952d138e60ee998c341c