Set haproxy global connection limit
Haproxy uses two separate maxconn values: the maxconn setting in the
global context is used to adjust file descriptor limits and allocate
resources, while the maxconn setting in the frontend context controls if
additional connections are accepted on that frontend.

This patch adds a maxconn to the global context equal to the sum of the
connection limit on each listener. For listeners which no connection
limit specified, we don't set this value and leave it at haproxy's
default (2000).

Change-Id: I30f3cf5f469db953a0002721a315f02694621b7d
Partial-Bug: #1544861