Set fullconn parameter in haproxy backends
By default haproxy sets the connection limit of any backend to 10% of
the connection limits for all its front-ends. For long-lasting but
low-load connections to a small number of back-end servers through the
load balancer, this means that the haproxy-based back-ends have an
effective connection limit that is much smaller than what the user
expects it to be.

The commit explicitly sets the connection limit for haproxy back-ends to
be equal to that of their front-ends.

Change-Id: Ie87722318fd2bc0005b089d74363c0812253c27f
Partial-Bug: #1622793