Remove last vestiges of oslo-incubator
This is a community goal for Ocata.  We're already not using any of
the code, but there are some other files and references left over.

This file used to be installed from oslo-incubator:
  tools/install_venv_common.py
It was meant to be used by the legacy run_tests.sh, which does not
exist in this repo.

Change-Id: I5040101102167ae3c2701019ceaad8bf8833a3c1
Partial-Bug: #1639103