Adds user_group option under haproxy_amphora
This fix adds the user_group configuration option under the haproxy_amphora
section, which is currently available in the neutron-lbaas code base.

The incentive for the above-mentioned addition is that in CentOS, Fedora and
RHEL based amphora images, the 'nogroup' user group does not exist by default.
Therefore users who wish to use those type of amphora images should have the
option to configure a different user group such as 'haproxy'.

The default value for  user group is 'nogroup' so the default behavior
remains intact.

Related-Bug #1548070
Change-Id: Ifac59889fa8120d974840bae2913587292f474c1