Clarify that octavia-db-manage should be used
It was not clear to users that we now use octavia-db-manage for
alembic actions in octavia.  Users would get an exception if they
tried to use native alembic due to changes to the env.py.
This patch updates the README.rst and env.py to clarify that
octavia-db-manage should be used.

Change-Id: I97e7fce5da167875694bc1a6abc69d47b3021487