Add common base type for v1 and v2
Since v1 and v2 types will share similar code, it makes sense for there
to be a base type for all types.  This also creates the types packages
under the v2 api package.

This also adds logic to allow for the renaming of fields from types to
data models.  For example: if the API accepts the field "admin_state_up"
but the data layer knows it as "enabled", then the type can define the
field "admin_state_up" and also define a mapping that says
"admin_state_up" will be mapped to "enabled" for data model
transformations.

Change-Id: I171e394daeb58ec90e9f531c91301c3dfd48dbb7