Fix Octavia v1 API project_id for POST
This patch corrects the project_id handling for POST calls to the
Octavia v1 API.  For load balancer create calls we use the specified
project_id if the user is an admin or noauth is specified.  If no
project_id is specified in the request we use the project_id from
the context.  If no project_id can be found we raise an exception.
For the other object POST methods we use the project_id
from the parent load balancer.

Change-Id: Ibf59541b8811e3bbe36cfec039f91e20036102e4
Closes-Bug: #1624145