Fix load balancer project_id handling on POST
This patch corrects the project_id handling for POST calls to the
Octavia v2 API for load balancers.  For load balancer create calls
we use the specified project_id if the user is an admin or noauth
is specified.  If no project_id is specified in the request we use
the project_id from the context.  If no project_id can be found we
raise an exception.

Change-Id: If0b1daf60cd422fa50fd2b4d9260a9859a862831