Tests helper function _send_requests should catch socket.error
Some tests, such as test_listener_basic for example, use helper function
_send_requests() to send traffic to a loadbalancer both in a positive
(expect to succeed) and negative (expect to fail) context.

In a case of a negative context,  _send_requests() currently won't catch
socket.error, which is a valid response from a loadbalancer in a case of
a request rejection.

F5's LBaaS driver replays with TCP RST in a case of request rejection and
they confirmed[1] that adding by catching socket.error in
_send_requests() tests work as expected.

[1] https://bugs.launchpad.net/octavia/+bug/1675702/comments/2

Change-Id: Iab001fc31ce81b2fd5e8e87333fa9a3391ecda99
Closes-Bug: #1675702