Fix an failure if the vip network has no gateway
If the VIP network was created in neutron with no gateway defined
the jinja template for the interfaces files would incorrectly put
"gateway None" into the interfaces file.  This causes a failure when
the interface is brought up.  This adds a check in the jinja template
to only add the gateway line if it is defined.

Change-Id: I3db02d61e34b767417853a4deae6ad84f64b64a9
Closes-Bug: #1687460