Fix an issue with failover on VIP net without DHCP
During a failover event, if the VIP subnet did not have DHCP
enabled the failover would fail.  The failover flow was not
refreshing the amphora data stored in the flow.  This caused
later tasks in the flow to incorrectly think that there was no
vrrp_ip assigned to the amphora.

This change adds an amphora refresh task to the flow to make sure
the flow data is updated from the database and the vrrp_ip is
available.  It also corrects a logging message to show the port ID.

Change-Id: Ibb9607a8aa14cdcfbaf5eca7b5206c586d22efce
Closes-Bug: #1694440
Closes-Bug: #1694426