Force get_all to only make one DB call
Previously our database repository "get_all" query would lazy load child
objects.  This means for each record returned, sqlalchemy will make multiple
additional calls to the database when we build the data model.
This patch changes the query to do a joined load so that all of the required
information for the data model is pulled over at one time.
This will increase the peak memory usage on the api server during the API
call, but it should still be at a reasonable level.

Change-Id: If83c5e8cd8d13d31ad0c5130e1b8b1a1209abadb