Fix a bad revert method and add hacking check
This patch fixes a revert method that was not handling extra parameters
being passed to it.
It also adds a hacking check to make sure this does not happen in the
future.
The patch also breaks the bad habit of compiling regex strings for every
line of code in the project.

Change-Id: If29e377204432e215bfea97f9d76bce0a442f4c8
(cherry picked from commit c3754dbf5a5ca6d2fa378cf95e30cd64fc8120a5)