Fix non-cascade LB delete with children
When the user attempts to delete a load balancer, without the cascade option,
the load balancer is locked in "PENDING_DELETE" prior to the check for child
objects. The child object check will return the error to the user, but would
leave the load balancer locked in "PENDING_DELETE".
This patch corrects the order to do the validation prior to locking the
load balancer in "PENDING_DELETE".

Change-Id: If1694743d098b1c1337601770a7b3466a43dda07
Story: 2001256
Task: 5786