Fix non-cascade LB delete with children
When the user attempts to delete a load balancer, without the cascade option,
the load balancer is locked in "PENDING_DELETE" prior to the check for child
objects. The child object check will return the error to the user, but would
leave the load balancer locked in "PENDING_DELETE".
This patch corrects the order to do the validation prior to locking the
load balancer in "PENDING_DELETE".

Story: 2001256
Task: 5786
(cherry picked from commit c05a8cfb88bd1219a086e8826f7783b64f1e8537)

Change-Id: I7364e6dbe6f647962df892e4c72cfdf7598c502b