Fix possible DB deadlock scenario
Under certain conditions if an error occurred during an API request the
lock session could be left open. This patch corrects that by extending the
try block to include all calls with the lock open.

The test case was creating a second pool on a listener, which results in
a 409 conflict error, but the lock was left open.

Change-Id: I81e08775c515602f315aa8da32ff342f96c4a676