Fix extracting nova fault string on build error
There was an edge case where if we were not able to extract nova interface
information our code that extracts the nova fault information would be
bypassed. This patch moves the fault extraction to a point where it will
be able to capture the fault string even if it was network interface
related.

Change-Id: Ia46522524e80022cd528e4902585213b1de42a44