Rename 'remove' to 'reset' in Quota API
The name of the remove_quota API call is misleading.
The real purpose of this call is to reset the quota
to its default value, not to remove it.

This patch is limited to the title of the call in v2,
and it does not modify the REST API in any way.

Task: 5868
Story: 2001295

Change-Id: Idfb1f3c3fe90d71434d0bb6a973e146ea3dea67c