Fix the failover API to not fail with immutable LB
A patch was added[1] that put an immutable check in the amphora failover
flow. However, when failover was triggered by the failover API, the load
balancer was already in PENDING_UPDATE. This caused the flow to fail and the
load balancer to be left in ERROR. This patch resolves those two code paths
to handle the load balancer locking the same.

[1] https://review.openstack.org/#/c/479109/

Change-Id: I6c8549c15a302add8c69ab95bde57fb2646455b6