Re-enable "Page Contents" and update user docs
This patch re-enables the page Table of Contents (ToC) for the Octavia
documentation. Previously the theme had the table of contents at the top
of each page which cluttered our index pages. Now the theme based ToC is
located on the left navigation menu.

This patch also reorganizes the user documentation section and adds a link
to our "Introduction to OpenStack Load Balancing" presentation from the
2017 Boston summit.

This change will not impact bookmarks.

Change-Id: Ib200f8b95e216b081d884da017fa1cce85f48763