Add deadlock retry wrapper for inital quota create
Since the initial quota record creation is using the autocommit session and
since it is creating the initial record, it will lock at the table(index)
level instead of at the record level. This can lead to a recoverable DB
Deadlock situation.
This was observed running the neutron-lbaas API tempest tests with concurrency
greater than 1.

This patch adds the oslo DB deadlock retry wrapper to this initial record
creation to allow deadlock recovery.

Change-Id: Ibbfa77f548fead3203cf081e203c5e70fe514111