Make keepalived initialization more predictable
This prevents forcing keepalived to renegotiate roles between amps more
times than necessary.

Keepalived is already running at the point that we do our initial "amp
start" command, so if we actually force a restart it loses its state and
has to renegotiate. I would consider the fact that this works without
issue right now "very good luck". Let's make this more sane, and only
reload if that's what we need to do.

Change-Id: Ie920bba9b2d718c59c6ce00c0f013058ed2634bb