Pin pip < 10 in the amphora image
pip 10 fails when installing the amphora-agent due to a package conflict
in oslo.config PyYAML and the python3-yaml installed in the base images.
We cannot backport the venv fix from master as the stable branch uses
the ubuntu element which does not have enough disk space to install with the
venv.
So to make a minimal fix for the stable branches we are going to pin pip
in the amphora image.

(cherry picked from commit 9e1ced71e1e270777b15adbcde8de03577375964)

Change-Id: I04bf04178c5cd169e3ff1a16a1c0423ef0a07b6d