Ignore a port not found when deleting an LB
There seems to be a case where we successfully retrieve the port
but when we try to update it might be gone because someone deleted
the vip port manually while our delete was scheduled.

Octavia does the right thing and marks the LB in ERROR bu this
will make our delete a bit more resilient to outside interference.

Change-Id: I847e46a674ba31c3e9669a99bf5c0268a3e26d23