Healthmanager shouldn't update NO_MONITOR members
During a create of a member on a pool with no healthmonitor, there is a
chance the member will be briefly updated to OFFLINE incorrectly.

The flow of events:
1. A Member is created in DB in NO_MONITOR.
2. A Member is passed to the handler.
3. A health message comes in, not including the member.
4. The Member is present in the DB but missing from the Amp, and is
   set to OFFLINE status.
5. The Member is created on the amp.
6. Another health messages comes in, and the Member is put back in the
   correct NO_MONITOR state.

Note: This patch temporarily makes the scenario gate non-voting as a
tempest change merged that tests two patches as one. The followon patch[1]
will re-enable it.

[1] https://review.openstack.org/#/c/612296/

Change-Id: I74c8cda0c7da88fedf652215630c701bb2761eef
(cherry picked from commit 267114fc90e2fa1e40114553cedd4fea4d2a8225)