Add tenant_id back to objects for backwards compat
To be 100% backwards compatible with neutron-lbaas, we need to continue
to return tenant_id in addition to project_id, even though it has been
deprecated and was actually removed previously from Octavia.

Adding it this way is much cleaner than adding it to every object and
will allow us to remove it very easily in the future.

Change-Id: I371a5bd51fc9acc4f6ef1e2b3b2fae9d4babd975