Providers: propose changes to API
While implementing an API driver for VMware loadbalancing platforms,
we ran into several difficulties which, I believe, might impact other
vendors as well.

The current approach of passing only the object id in delete operation
forces the driver to either run within Octavia's context, or maintain
a full state in order to map a sub-object id (e.g listener), to a
loadbalancer, which it should be deleted from.

With updates the problem is more complex as for reference the driver
has to maintain a full state of the object - not only mapping to the
loadbalancer. We already know what maintaining a mirror state is painful
as we experienced this for a while with LBaaSv2 and Octavia.
As an alternative, the driver could retreive the properties from the
appliance which may be a costly operation and will only resolve the update
challenges.

Change-Id: Ide4beaf73bb99be5c3b4c71c12fecef3499f45d9