Providers: propose changes to create pool API
Currently the driver data model for pool does not contain the
listener_id.
Unless we add it, there is no other way for the drivers to know which
listener was used since the DB session was not yet committed.
This patch suggests to add the listener_id to the object that the
driver receives.

Change-Id: If489f7bb956381bc5223556d118cb48571f0873f