Allow using spaces for primary common name in SSL certificates
The common name is used as a file name inside the HAproxy
configuration file.  However, a common name can include spaces
and it will result in a configuration file that simply doesn't
work because of the spaces.

The patch changes the functionality so that it instead creates
a SHA1 hash of the certificate and uses that as the file name
to avoid those issues.

Change-Id: I039ed0b40df8b72a1238f8896548fe77086c530c
(cherry picked from commit 852191feb22f0968c4338c06952ba88191df8087)