Updates the amphora driver for new commit model
The provider driver model changes the way updates are committed
to the database. This patch updates the amphora driver to work with
the new update model for provider drivers. Specificially that the
objects the amphora driver are using will already be updated.

This patch also corrects a bug in the batch member update that would
not update members correctly for provider drivers.

Change-Id: I03ba4d7755bbb0c034c75f29f4a59b1b708daac5