Translate expected package names to installed ones
This patch fixes an issue that caused an internal error to the amphora
agent and led amphorae to transition to FAILURE state. The issue was
that on CentOS-based amphorae HAproxy package name is 'haproxy18' while
the expected package name is 'haproxy'.

The patch introduces a package name per distro map that is consulted
whenever a package has a different name than the one expected. If
there's a miss, it falls back to the expected package name.

Story: 2002958
Task: 22960

Change-Id: I59dfa3f9bd5974302ec72343444813232ec2ea02