Temporarily disable flow diagraming
The networkx package changed their dependecy model in version 2.1 and
now uses "pydot" and adds the dependency via setuptools "extras".
Unfortunately they only define an "extra" of "all" which installs packages
we do not need and fail to install.

So until the "extras" are fixed in networkx, we need to disable flow
diagraming.

Change-Id: I4f9082c267dc7da14b81cb551eeb10bbf58c175a
See: https://github.com/networkx/networkx/pull/3080