"Resolve" bandit issue with sha1 hashes
We only use them as a consistent hash, not for anything security
related. Even if someone could manage a hash collision or something,
the worst that could happen is their own LB stops working...
So, just tell bandit to ignore them.
Added a docstring to one function to attempt to be extra clear.

Change-Id: Ic337c32020e12183e2246b844ded07a83c6d3ad8