Fix a lifecycle bug with child objects
In some edge cases, when a request comes in to update a child
resource (member, HM, or l7rule), the API would not always make
sure the parent object (pool or l7policy) was in a mutable status.

This patch fixes this by checking the parent provisioning_status
before allowing a muttable change on a child object.

It also resolves two cases where a parent object status was not being
reset on a reverted flow.

Change-Id: I86f76bebce0993215fd34ccd33251ba1e6c325a9
Story: 2005249
Task: 30043